WIP: continuing to fix tests
