--- crates/learner/tests/database/mod.rs
@@ -1,3 +1,0 @@
-use super::*;
-
-mod instruction;
