+++ crates/learner/tests/llm/mod.rs
@@ -1,3 +1,5 @@
+use learner::database::Add;
+
 use super::*;
 
 #[ignore = "Can't run this in general -- relies on local LLM endpoint."]
@@ -5,18 +7,22 @@ use super::*;
 #[traced_test]
 async fn test_download_then_send_pdf() -> Result<(), Box<dyn Error>> {
   // Download a PDF
-  let dir = tempdir().unwrap();
-  let paper = Paper::new("https://eprint.iacr.org/2016/260").await.unwrap();
-  paper.download_pdf(dir.path()).await.unwrap();
+  let (mut learner, _cfg_dir, _db_dir, _strg_dir) = create_test_learner().await;
+  let paper = learner.retriever.get_paper("https://eprint.iacr.org/2016/260").await?;
+  // let paper = Paper::new("https://eprint.iacr.org/2016/260").await.unwrap();
+
+  // paper.download_pdf(dir.path()).await.unwrap();
+  Add::complete(&paper).execute(&mut learner.database).await?;
 
   // Get the content of the PDF
-  let formatted_title = format::format_title(&paper.title, None); // use default 50
-  let path = dir.into_path().join(format!("{}.pdf", formatted_title));
-  let pdf_content = PDFContentBuilder::new().path(path).analyze().unwrap();
 
-  let mut message = "Please act like a researcher and digest this text from a PDF for me and give \
-                     me an excellent summary. The summary can be long and descriptive. \n"
-    .to_owned();
+  let path = learner.database.get_storage_path().await?.join(paper.filename());
+  let pdf_content = PDFContentBuilder::new().path(path).analyze()?;
+
+  let mut message =
+    "Please act like a researcher and digest this text from a PDF for me and give  me an \
+     excellent summary. The summary can be long and descriptive. \n"
+      .to_owned();
 
   message.push_str(&serde_json::to_string(&pdf_content.metadata).unwrap());
   message.push_str(&serde_json::to_string(&pdf_content.pages[0..5]).unwrap());
